+++ tests/test_plugin_webex_teams.py
@@ -63,6 +63,15 @@ apprise_url_tests = (
         },
     ),
     (
+        "wxteams://?token={}".format("a" * 80),
+        {
+            # token provided - we're good
+            "instance": NotifyWebexTeams,
+            # Our expected url(privacy=True) startswith() response:
+            "privacy_url": "wxteams://a...a/",
+        },
+    ),
+    (
         "webex://{}".format("a" * 140),
         {
             # token provided - we're good
